+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/verticalmode/PaymentMethodRowDisclosureButtonScreenshotTest.kt
@@ -144,16 +144,25 @@ internal class PaymentMethodRowDisclosureButtonScreenshotTest {
 
     @OptIn(AppearanceAPIAdditionsPreview::class)
     @Test
-    fun testTitleFont() {
+    fun testFonts() {
         testPaymentMethodRowButton_Disclosure(
+            subtitle = "this is a subtitle",
             appearance = PaymentSheet.Appearance.Embedded.Builder()
                 .rowStyle(FlatWithDisclosure.default)
                 .titleFont(
                     PaymentSheet.Typography.Font(
                         fontFamily = com.stripe.android.paymentsheet.R.font.cursive,
                         fontSizeSp = 20f,
-                        fontWeight = 30,
+                        fontWeight = 500,
                         letterSpacingSp = 10f
+                    )
+                )
+                .subtitleFont(
+                    PaymentSheet.Typography.Font(
+                        fontFamily = com.stripe.android.paymentsheet.R.font.cursive,
+                        fontSizeSp = 12f,
+                        fontWeight = 200,
+                        letterSpacingSp = 5f
                     )
                 )
                 .build()