+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## XX.XX.XX - 20XX-XX-XX
 ### PaymentSheet
 * [FIXED][11184](https://github.com/stripe/stripe-android/pull/11184) PaymentSheet will now accurately show a success state when confirming Blik and UPI PaymentIntents using a Stripe Connect account ID.
+* [ADDED][11205](https://github.com/stripe/stripe-android/pull/11205) Added `subtitleFont` to `PaymentSheet.Appearance.Embedded` for customization of EmbeddedPaymentElement payment method row subtitles.
 
 ## 21.21.0 - 2025-07-21
 ### PaymentSheet